safari-600.7-branch/LayoutTests/platform/mac-wk2/accessibility/content-inset-scrollview-frame.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that a page with a content inset will not affect the bounds of the main frame scroll view. It will appear as the same as the web area's sbounds");

    if (window.accessibilityController) {
        var webArea = accessibilityController.rootElement.childAtIndex(0);
        var webX = webArea.x;
        var webY = webArea.y;

        var scrollView = webArea.parentElement();
        var scrollViewX = scrollView.x;
        var scrollViewY = scrollView.y;
  
        debug("The position of the web area and the scroll view should be the same when there's no content inset");
        shouldBeTrue("webX == scrollViewX && webY == scrollViewY");
 
        window.internals.setTopContentInset(100);

        debug("After setting the content inset, the new y position should be less than the initial");
        shouldBe("scrollViewY - scrollView.y", "100")
    }
</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
